refactor(writers): share the identical markdown-family block helpers
